requirements.txt
PyYAML>=6.0
tqdm>=4.66
numpy>=1.26
matplotlib>=3.8
fastapi>=0.115.0
uvicorn>=0.32.0

openai>=1.0.0
python-dotenv>=1.0.0

pytest>=7.4.0
pytest-asyncio>=0.21.0
